refactor: Separate logic from service module
The service module has been created to separate the services that the API offers from the logic JS file
